Speed module pings
Also other stuff, like reorganization, memory game changes, and removed useless file
