+++ api/mkwii/mkwii_file_handling.py
@@ -27,7 +27,21 @@ async def handle_mkwii_files(message, attachments, file_dict, self):
                     return
 
                 if await is_time_over(message.author.id):
-                    await message.channel.send("You can't submit, your time is up!")
+                    # If they have not submitted, they get a different message.
+                    async with get_session() as session:
+                        query = select(Submissions.user_id).where(Submissions.user_id == message.author.id)
+                        result = (await session.execute(query)).first()
+
+
+                        if result is None:
+                            message_to_send = (f"You can't submit, your time is up! If you wish to send in a late submission, "
+                                       f"please DM the current host so they can add your submission manually.")
+
+                        else:
+                            message_to_send = "You can't submit, your time is up!"
+
+
+                    await message.channel.send(message_to_send)
                     return
 
             ##################################################
