+++ commands/fun/memory.py
@@ -45,21 +45,17 @@ class Memory(commands.Cog):
 
     @commands.hybrid_command(name="memory", description="Play a memory game!", with_app_command=True)
     async def command(self, ctx, size: int = 4):
-        if size % 2 != 0:
-            await ctx.send("Board size must be an even number!")
-            return
-
         if ctx.author.id in self.ongoing_games:
             await ctx.send("You're already playing a game!")
             return
 
-        lives = int(size / 2 + 1)
+        lives = int(size - 1)
 
         # Create the embed for game instructions
         embed = discord.Embed(
             title="Memory Game Instructions",
             description=f"Welcome to the memory game!\n\n"
-                        "In this game, a board is presented to you and need to match all pairs of emojies to win! "
+                        "In this game, a board is presented to you and you need to match all pairs of emojis to win! "
                         "To select a tile, you need to send a message in the following format: row, column\n",
             color=discord.Color.blue()
         )
@@ -98,18 +94,34 @@ class Memory(commands.Cog):
         try:
             # Get list of emojis from the server
             guild_emojis = ctx.guild.emojis
-            if len(guild_emojis) < size * size // 2:
+            num_emojis_needed = (size * size - 1) // 2 if size % 2 != 0 else (size * size) // 2
+            if len(guild_emojis) < num_emojis_needed:
                 await ctx.send(f"Not enough emojis on the server for a {size}x{size} board!")
                 return
 
             # Prepare unique pairs of emojis and shuffle the board
-            unique_emojis = random.sample(guild_emojis, size * size // 2)
+            unique_emojis = random.sample(guild_emojis, num_emojis_needed)
             board_emojis = unique_emojis * 2
             random.shuffle(board_emojis)
 
             # Create hidden board with ? placeholders
             hidden_board = [['❓' for _ in range(size)] for _ in range(size)]
-            display_board = [[board_emojis.pop() for _ in range(size)] for _ in range(size)]
+
+            # Handle odd-sized board middle tile (❌)
+            middle_pos = (size // 2, size // 2) if size % 2 != 0 else None
+
+            # Create the display board and leave the middle tile as '❌' if board size is odd
+            display_board = [
+                [
+                    board_emojis.pop() if middle_pos is None or (i, j) != middle_pos else '❌'
+                    for j in range(size)
+                ]
+                for i in range(size)
+            ]
+
+            # Make sure the middle tile in hidden_board is set to '❌' if the board size is odd
+            if middle_pos:
+                hidden_board[middle_pos[0]][middle_pos[1]] = '❌'
 
             # Function to display the board in a formatted way without row numbers
             def format_board(board):
@@ -117,10 +129,13 @@ class Memory(commands.Cog):
 
             # Function to reveal a 3x3 area (Fog of War effect) around the selected tile
             def reveal_area(row, col, radius=1):
-                revealed_board = [['❓' for _ in range(size)] for _ in range(size)]
+                revealed_board = [
+                    ['❓' if middle_pos is None or (r, c) != middle_pos else '❌' for c in range(size)]
+                    for r in range(size)
+                ]
                 for r in range(max(0, row - radius), min(size, row + radius + 1)):
                     for c in range(max(0, col - radius), min(size, col + radius + 1)):
-                        if hidden_board[r][c] == '❓':  # Only reveal hidden tiles
+                        if hidden_board[r][c] == '❓' and (r, c) != middle_pos:  # Only reveal hidden tiles, skip middle
                             revealed_board[r][c] = display_board[r][c]
                         else:
                             revealed_board[r][c] = hidden_board[r][c]  # Show already revealed/matched tiles
@@ -133,10 +148,13 @@ class Memory(commands.Cog):
 
             # Function to keep only the selected tile visible, hiding the surroundings
             def keep_only_selected_visible(row, col):
-                single_tile_board = [['❓' for _ in range(size)] for _ in range(size)]
+                single_tile_board = [
+                    ['❓' if middle_pos is None or (r, c) != middle_pos else '❌' for c in range(size)]
+                    for r in range(size)
+                ]
                 for r in range(size):
                     for c in range(size):
-                        if hidden_board[r][c] != '❓':  # Keep matched pairs visible
+                        if hidden_board[r][c] != '❓' or (r, c) == middle_pos:  # Keep matched pairs visible and middle 'X'
                             single_tile_board[r][c] = hidden_board[r][c]
                 # Keep only the selected tile visible
                 single_tile_board[row][col] = display_board[row][col]
@@ -152,16 +170,19 @@ class Memory(commands.Cog):
                 # Check if the input is in the format "row,column" and within bounds (1-based)
                 try:
                     row, col = map(int, choice.split(","))
+                    # Ensure the player isn't selecting the middle "X" tile
+                    if middle_pos and row == middle_pos[0] + 1 and col == middle_pos[1] + 1:
+                        return False
                     return 1 <= row <= size and 1 <= col <= size and hidden_board[row - 1][col - 1] == '❓'
                 except (ValueError, IndexError):
                     return False
 
             # Main game loop with timeout handling
-            while lives > 0 and len(found_pairs) < size * size:
+            while lives > 0 and len(found_pairs) < size * size - (1 if middle_pos else 0):  # Adjust for missing middle tile
                 try:
-                    # Get first choice from the user with a 60-second timeout
+                    # Get first choice from the user with a 40-second timeout
                     await ctx.send(f"Select first tile! Lives remaining: {lives}")
-                    msg = await self.bot.wait_for('message', timeout=60.0, check=lambda m: m.author == ctx.author and is_valid_choice(m.content))
+                    msg = await self.bot.wait_for('message', timeout=40.0, check=lambda m: m.author == ctx.author and is_valid_choice(m.content))
                     row1, col1 = map(int, msg.content.split(","))
                     row1 -= 1  # Adjust for 0-based indexing
                     col1 -= 1
@@ -175,9 +196,9 @@ class Memory(commands.Cog):
                     visible_only_selected = keep_only_selected_visible(row1, col1)
                     await board_message.edit(content=format_board(visible_only_selected))  # Edit instead of delete
 
-                    # Get second choice from the user with a 60-second timeout
+                    # Get second choice from the user with a 40-second timeout
                     await ctx.send(f"Select second tile! Lives remaining: {lives}")
-                    msg = await self.bot.wait_for('message', timeout=60.0, check=lambda m: m.author == ctx.author and is_valid_choice(m.content))
+                    msg = await self.bot.wait_for('message', timeout=40.0, check=lambda m: m.author == ctx.author and is_valid_choice(m.content))
                     row2, col2 = map(int, msg.content.split(","))
                     row2 -= 1  # Adjust for 0-based indexing
                     col2 -= 1
@@ -215,7 +236,7 @@ class Memory(commands.Cog):
             # End game messages
             if lives == 0:
                 await ctx.send("Game over! You're out of lives.")
-            elif len(found_pairs) == size * size:
+            elif len(found_pairs) == size * size - (1 if middle_pos else 0):  # Adjust for missing middle tile
                 await ctx.send("Congratulations! You found all pairs!")
             else:
                 await ctx.send("The game has ended due to inactivity.")
